# src/main/java/net/umi/tarification_itn/model/dto/RefreshTokenRequestDTO.java
package net.umi.tarification_itn.model.dto;

import jakarta.validation.constraints.NotBlank;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class RefreshTokenRequestDTO {
    @NotBlank(message = "Le refresh token est requis")
    private String refreshToken;
}